requirements.txt
streamlit==1.24.0
googletrans==3.1.0a0
gTTS==2.3.2
deep-translator==1.11.4
transformers==4.30.2
torch==2.0.1
indic-nlp-library==0.92
python-docx==0.8.11
PyPDF2==3.0.1
nltk==3.8.1
spacy==3.5.3
sounddevice==0.4.6
soundfile==0.12.1
pillow==9.5.0
